+++ tests/vali_tests/test_positions_penalty.py
@@ -39,11 +39,10 @@ class TestPositionsPenalty(TestBase):
         )
         self.mock_metagraph = MockMetagraph([self.DEFAULT_MINER_HOTKEY])
         self.position_manager = PositionManager(metagraph=self.mock_metagraph, running_unit_tests=True)
-        self.position_manager.init_cache_files()
-        self.position_manager.clear_all_miner_positions_from_disk()
+        self.position_manager.clear_all_miner_positions()
 
     def tearDown(self):
-        self.position_manager.clear_all_miner_positions_from_disk()
+        self.position_manager.clear_all_miner_positions()
         super().tearDown()
 
     def test_max_positional_return_ratio(self):
