+++ tests/vali_tests/test_time_util.py
@@ -37,8 +37,7 @@ class TestTimeUtil(TestBase):
         )
         self.mock_metagraph = MockMetagraph([self.DEFAULT_MINER_HOTKEY])
         self.position_manager = PositionManager(metagraph=self.mock_metagraph, running_unit_tests=True)
-        self.position_manager.init_cache_files()
-        self.position_manager.clear_all_miner_positions_from_disk()
+        self.position_manager.clear_all_miner_positions()
 
     def test_n_crypto_intervals(self):
         prev_delta = None
