+++ vali_objects/utils/auto_sync.py
@@ -13,8 +13,10 @@ import bittensor as bt
 
 
 class PositionSyncer(ValidatorSyncBase):
-    def __init__(self, shutdown_dict=None, signal_sync_lock=None, signal_sync_condition=None, n_orders_being_processed=None, running_unit_tests=False):
-        super().__init__(shutdown_dict, signal_sync_lock, signal_sync_condition, n_orders_being_processed, running_unit_tests=running_unit_tests)
+    def __init__(self, shutdown_dict=None, signal_sync_lock=None, signal_sync_condition=None,
+                 n_orders_being_processed=None, running_unit_tests=False, position_manager=None):
+        super().__init__(shutdown_dict, signal_sync_lock, signal_sync_condition, n_orders_being_processed,
+                         running_unit_tests=running_unit_tests, position_manager=position_manager)
         self.force_ran_on_boot = True
 
     def read_validator_checkpoint_from_gcloud_zip(url):
